atm_banking/requirements.txt
Flask == 2.3.2
mysql_connector-python == 8.0.33
